LayoutTests/imported/w3c/web-platform-tests/content-security-policy/wasm-unsafe-eval/default-src-blocks-wasm.any-expected.txt


PASS WebAssembly.instantiate() is blocked
PASS WebAssembly.compile() is blocked
PASS new WebAssembly.Module() is blocked
